Integrate flatten_chapters.py
Books with nested chapters will not be parsed by the jq filter
unless the chapters in the extra_chapter_file json is flattened.
The flatten_chapters.py python script was written by
https://github.com/amz71 in Issue #224
https://github.com/KrumpetPirate/AAXtoMP3/issues/224

The next step is assimilating this new script/feature to the rest
of the script, such as the careful method for checking that
dependencies (python in this case) exist on the user's machine,
and always calling external dependencies with variables, rather
than by name directly, to ensure portability.