Allow outer loop to process next aax file
There is an error occurring within the final cover art block that is
causing the script to exit early. Even with errexit flag off. Not sure
yet what the error is but this block to add cover art seems to be only
needed for audiobooks that were not split into chapter files. If the
book was split into chapters, the output_file they were split from was
removed, so checking for the output_file skips this block and allows
the outer for loop to process the next aax file supplied to the script
via positional args.